backend/requirements.txt
 fastapi
 openai
 qdrant-client
 python-dotenv
 uvicorn
 httpx
 pydantic
 cohere